chore: reestructuracion paquete de modelos a model
