+++ .idea/workspace.xml
@@ -5,11 +5,34 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="6553db53-e3fe-47a3-a420-6c14b605080a" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/gestion/GestorReserva.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/interfaces/IDiaDeSol.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/interfaces/IReserva.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/model/alojamiento/Alojamiento.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/model/alojamiento/Apartamento.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/model/alojamiento/DiaDeSol.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/model/alojamiento/Finca.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/model/alojamiento/Habitacion.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/model/alojamiento/Hotel.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/model/data/HuespedData.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/model/data/ReservaData.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/Main.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/gestion/FormularioHelper.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/gestion/FormularioHelper.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/gestion/GestorMenus.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/gestion/GestorMenus.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/logicaNegocio/FiltroDeHabitacion.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/logicaNegocio/FiltroDeHabitacion.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Alojamiento.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Alojamiento.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Finca.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Finca.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Hotel.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Hotel.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/logicaNegocio/ReservaImplementation.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/bookinghotels/logicaNegocio/ReservaImplementation.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Alojamiento.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Apartamento.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/ClienteData.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/DiaDeSolData.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Finca.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Habitacion.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/Hotel.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/IDiaDeSol.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/IReserva.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/ReservaData.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/bookinghotels/modelos/dto/ParametrosBusqueda.java" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
