+++ src/main/java/com/bookinghotels/interfaces/IDiaDeSol.java
@@ -1,0 +1,9 @@
+package com.bookinghotels.interfaces;
+
+import java.time.LocalDate;
+
+public interface IDiaDeSol {
+    boolean tieneDiaDeSol();
+    void mostrarInfoDiaDeSol(Integer cantPersonas, LocalDate fechaInicio);
+    float calcularPrecioBaseDiaSol(Integer cantPersonas);
+}
