+++ src/main/java/com/bookinghotels/model/data/HuespedData.java
@@ -1,0 +1,71 @@
+package com.bookinghotels.model.data;
+
+import java.time.LocalDate;
+
+public class HuespedData {
+    private String nombre;
+    private String apellido;
+    private LocalDate fechaNacimiento;
+    private String numeroTel;
+    private String correo;
+    private String nacionalidad;
+
+    // Constructor
+    public HuespedData(String nombre, String apellido, LocalDate fechaNacimiento, String numeroTel, String correo, String nacionalidad) {
+        this.nombre = nombre;
+        this.apellido = apellido;
+        this.fechaNacimiento = fechaNacimiento;
+        this.numeroTel = numeroTel;
+        this.correo = correo;
+        this.nacionalidad = nacionalidad;
+    }
+
+    // Getters and Setters
+    public String getNombre() {
+        return nombre;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+    public String getApellido() {
+        return apellido;
+    }
+
+    public void setApellido(String apellido) {
+        this.apellido = apellido;
+    }
+
+    public LocalDate getFechaNacimiento() {
+        return fechaNacimiento;
+    }
+
+    public void setFechaNacimiento(LocalDate fechaNacimiento) {
+        this.fechaNacimiento = fechaNacimiento;
+    }
+
+    public String getNumeroTel() {
+        return numeroTel;
+    }
+
+    public void setNumeroTel(String numeroTel) {
+        this.numeroTel = numeroTel;
+    }
+
+    public String getCorreo() {
+        return correo;
+    }
+
+    public void setCorreo(String correo) {
+        this.correo = correo;
+    }
+
+    public String getNacionalidad() {
+        return nacionalidad;
+    }
+
+    public void setNacionalidad(String nacionalidad) {
+        this.nacionalidad = nacionalidad;
+    }
+}
