--- src/main/java/com/bookinghotels/modelos/IDiaDeSol.java
@@ -1,9 +1,0 @@
-package com.bookinghotels.modelos;
-
-import java.time.LocalDate;
-
-public interface IDiaDeSol {
-     boolean tieneDiaDeSol();
-     void mostrarInfoDiaDeSol(Integer cantPersonas, LocalDate fechaInicio);
-     float calcularPrecioBaseDiaSol(Integer cantPersonas);
-}
